pyproject.toml
[tool.poetry]
name = "pste"
version = "0.3.0"
description = "Just a simple file hosting application inspired by the likes of pomf.se and teknik.io."
authors = ["Lee Watson <rev@revthefox.co.uk>"]
license = "GPL-3.0+"

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.0"
flask-login = "^0.5.0"
flask-sqlalchemy = "^2.4.0"
flask-migrate = "^2.5.0"
flask-wtf = "^0.14"
flask-assets = "^2.0"
libsass = "^0.20.0"
cssmin = "^0.2.0"
dynaconf = { version = "^3.0.0", extras = ["yaml"] }
passlib = { extras = ["argon2", "bcrypt"], version = "^1.7.0" }
python-magic = "^0.4.0"
humanize = "^3.0.0"
pygments = "^2.6.0"
sentry-sdk = { extras = ["flask"], version = "^0.18.0" }
psycopg2 = { version = "^2.8.0", optional = true }
mysqlclient = { version = "^2.0", optional = true }
gunicorn = { version = "^20.0", optional = true, extras = ["gevent"] }
email_validator = "^1.0.5"
python-dotenv = "^0.14.0"
python-box = "^5.1.1"
pyyaml = "^5.3.1"

[tool.poetry.dev-dependencies]
pre-commit = "^2.7.1"
flake8 = "^3.8.3"
black = "^20.8b1"

[tool.poetry.extras]
mysql = ["mysqlclient"]
postgresql = ["psycopg2"]
gunicorn = ["gunicorn"]

[tool.black]
exclude = '''
(
    pste/storage/|
    pste/migrations/
)
'''

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
